+++ src/lightning_hydra_zen_template/data/mnist.py
