+++ tests/data/test_mnist.py
@@ -1,7 +1,7 @@
 import pytest
 from rootutils import find_root
 
-from lightning_hydra_zen_template.datamodule.mnist import MNISTDataModule
+from lightning_hydra_zen_template.data.mnist import MNISTDataModule
 
 ROOT_DIR = find_root()
 DATA_DIR = ROOT_DIR / "data" / "raw"
